feat: modify test for todo service and todo controller

+++ todo-list-be/src/todo/todo.controller.ts
@@ -1,4 +1,13 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import {
+  BadRequestException,
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Patch,
+  Post,
+} from '@nestjs/common';
 import { TodoService } from './todo.service';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
@@ -19,16 +28,46 @@ export class TodoController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.todoService.findOne(+id);
+    if (isNaN(+id)) {
+      throw new BadRequestException('Invalid ID');
+    }
+
+    const todo = this.todoService.findOne(+id);
+
+    if (!todo) {
+      throw new BadRequestException('Todo not found');
+    }
+
+    return todo;
   }
 
   @Patch(':id')
   update(@Param('id') id: string, @Body() updateTodoDto: UpdateTodoDto) {
-    return this.todoService.update(+id, updateTodoDto);
+    if (isNaN(+id)) {
+      throw new BadRequestException('Invalid ID');
+    }
+
+    const updatedTodo = this.todoService.update(+id, updateTodoDto);
+
+    if (!updatedTodo) {
+      throw new BadRequestException('Todo not found');
+    }
+
+    return updatedTodo;
   }
 
   @Delete(':id')
   remove(@Param('id') id: string) {
-    return this.todoService.remove(+id);
+    if (isNaN(+id)) {
+      throw new BadRequestException('Invalid ID');
+    }
+
+    const removedTodo = this.todoService.remove(+id);
+
+    if (!removedTodo) {
+      throw new BadRequestException('Todo not found');
+    }
+
+    return removedTodo;
   }
 }
